requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
idna==2.8
itsdangerous==1.1.0
Jinja2>=2.10.1
MarkupSafe==1.1.1
Pillow==6.0.0
pyserial==3.4
requests==2.22.0
six==1.12.0
urllib3==1.25.3
waitress==1.2.1
Werkzeug==0.14.1
